feat: Add scaleout coffea notebook
